move thread messaging calls to separate files
