+++ blakserv/async.c
@@ -302,10 +302,10 @@ Bool CheckMaintenanceMask(SOCKADDR_IN6 *addr,int len_addr)
 			if (mask.u.Byte[k] != 0 && mask.u.Byte[k] != addr->sin6_addr.u.Byte[k])
 			{
 #else
-                for (int k = 0; k < sizeof(mask.s6_addr); k++)
-                {
-                        if (mask.s6_addr[k] != 0 && mask.s6_addr[k] != addr->sin6_addr.s6_addr[k])
-                        {
+        for (int k = 0; k < sizeof(mask.s6_addr); k++)
+        {
+            if (mask.s6_addr[k] != 0 && mask.s6_addr[k] != addr->sin6_addr.s6_addr[k])
+            {
 #endif
 				// mismatch
 				skip = 1;
