+++ blakserv/makefile.linux
@@ -94,9 +94,9 @@ OBJS =  \
 	$(OUTDIR)/files.obj \
 	$(OUTDIR)/sprocket.obj \
 	$(OUTDIR)/linux-critical_section.obj \
-	$(OUTDIR)/linux-threading.obj \
 	$(OUTDIR)/interface_linux.obj \
 	$(OUTDIR)/mutex_linux.obj \
+    $(OUTDIR)/tmsgloop_linux.obj \
 
 
 all : makedirs $(OUTDIR)/blakserv
