+++ blakserv/timer.c
@@ -135,7 +135,9 @@ void AddTimerNode(timer_node *t)
 
       /* we're making a new first-timer, so the time main loop should wait might
 	 have changed, so have it break out of loop and recalibrate */
-      PostThreadMessage(main_thread_id,WM_BLAK_MAIN_RECALIBRATE,0,0);
+
+      MessagePost(main_thread_id,WM_BLAK_MAIN_RECALIBRATE,0,0);
+
       return;
    }
 
@@ -245,12 +247,12 @@ Bool DeleteTimer(int timer_id)
    {
       if (t->timer_id == timer_id)
       {
-	 prev->next = t->next;
+         prev->next = t->next;
 
-	 /* put deleted timer on deleted_timer list */
-	 StoreDeletedTimer(t);
+         /* put deleted timer on deleted_timer list */
+         StoreDeletedTimer(t);
 
-	 return True;
+         return True;
       }
       prev = t;
       t = t->next;
@@ -268,6 +270,7 @@ Bool DeleteTimer(int timer_id)
    }
    dprintf("\n");
 #endif
+
    return False;
 }
 
@@ -286,11 +289,11 @@ void TimerActivate()
    now = GetMilliCount();
    if (now > timers->time)
    {
-	/*
+   /*
      if (now - timers->time > TIMER_DELAY_WARN)
        dprintf("Timer handled %i.%03is late\n",
-	       (now-timers->time)/1000,(now-timers->time)%1000);
-	*/
+         (now-timers->time)/1000,(now-timers->time)%1000);
+   */
 
       object_id = timers->object_id;
       message_id = timers->message_id;
@@ -319,7 +322,6 @@ Bool InMainLoop(void)
 
 void ServiceTimers(void)
 {
-#ifdef BLAK_PLATFORM_WINDOWS
    MSG msg;
    INT64 ms;
 
@@ -332,172 +334,86 @@ void ServiceTimers(void)
    eprintf("-------------------------------------------------------------------------------------\n");
 
    in_main_loop = True;
+
+#ifdef BLAK_PLATFORM_WINDOWS
    SetWindowText(hwndMain, ConfigStr(CONSOLE_CAPTION));
+#endif
 
    AsyncSocketStart();
 
    for(;;)
    {
       if (timers == NULL)
-			ms = 500;
+         ms = 500;
       else
       {
-			ms = timers->time - GetMilliCount();
-			if (ms <= 0)
-				ms = 0;
-	 
-			if (ms > 500)
-				ms = 500;
+         ms = timers->time - GetMilliCount();
+         if (ms <= 0)
+            ms = 0;
+
+         if (ms > 500)
+            ms = 500;
       }	 
       
-      if (MsgWaitForMultipleObjects(0,NULL,0,(DWORD)ms,QS_ALLINPUT) == WAIT_OBJECT_0)
+      if (WaitForAnyMessageWithTimeout(ms))
       {
-	 while (PeekMessage(&msg,NULL,0,0,PM_REMOVE))
-	 {
-	    if (msg.message == WM_QUIT)
-	    {
-	       lprintf("ServiceTimers shutting down the server\n");   
-	       return;
-	    }
+         while (MessagePeek(&msg,NULL,0,0,PM_REMOVE))
+         {
+            if (msg.message == WM_QUIT)
+            {
+               lprintf("ServiceTimers shutting down the server\n");   
+               return;
+            }
 	    
-	    switch (msg.message)
-	    {
-	    case WM_BLAK_MAIN_READ :
-	       EnterServerLock();
-	       
-	       PollSession(msg.lParam);
-	       TimerActivate();
-	       
-	       LeaveServerLock();
-	       break;
-	    case WM_BLAK_MAIN_RECALIBRATE :
-	       /* new soonest timer, so we should recalculate our time left... 
-		  so we just need to restart the loop! */
-	       break;
+            switch (msg.message)
+            {
+               case WM_BLAK_MAIN_READ :
+               EnterServerLock();
 
-	    case WM_BLAK_MAIN_DELETE_ACCOUNT :
-	       EnterServerLock();
-	       DeleteAccountAndAssociatedUsersByID(msg.lParam);
-	       LeaveServerLock();
-	       break;
+               PollSession(msg.lParam);
+               TimerActivate();
 
-	    case WM_BLAK_MAIN_VERIFIED_LOGIN :
-	       EnterServerLock();
-	       VerifiedLoginSession(msg.lParam);
-	       LeaveServerLock();
-	       break;
+               LeaveServerLock();
+               break;
 
-        case WM_BLAK_MAIN_LOAD_GAME :
-          EnterServerLock();
-          LoadFromKod(msg.lParam);
-          LeaveServerLock();
-          break;
+               case WM_BLAK_MAIN_RECALIBRATE :
+               /* new soonest timer, so we should recalculate our time left... 
+               so we just need to restart the loop! */
+               break;
 
-	    default :
-	       dprintf("ServiceTimers got unknown message %i\n",msg.message);
-	       break;
-	    }
-	 }
+               case WM_BLAK_MAIN_DELETE_ACCOUNT :
+               EnterServerLock();
+               DeleteAccountAndAssociatedUsersByID(msg.lParam);
+               LeaveServerLock();
+               break;
+
+               case WM_BLAK_MAIN_VERIFIED_LOGIN :
+               EnterServerLock();
+               VerifiedLoginSession(msg.lParam);
+               LeaveServerLock();
+               break;
+
+               case WM_BLAK_MAIN_LOAD_GAME :
+               EnterServerLock();
+               LoadFromKod(msg.lParam);
+               LeaveServerLock();
+               break;
+
+               default :
+               dprintf("ServiceTimers got unknown message %i\n",msg.message);
+               break;
+            }
+         }
       }
       else
       {
-	 /* a Blakod timer is ready to go */
-	 
-	 EnterServerLock();
-	 PollSessions(); /* really just need to check session timers */
-	 TimerActivate();
-	 LeaveServerLock();
+         /* a Blakod timer is ready to go */
+         EnterServerLock();
+         PollSessions(); /* really just need to check session timers */
+         TimerActivate();
+         LeaveServerLock();
       }
    }
-#else
-    MSG msg;
-    INT64 ms;
-
-    lprintf("Status: %i accounts\n",GetNextAccountID());
-
-    lprintf("-------------------------------------------------------------------------------------\n");
-    dprintf("-------------------------------------------------------------------------------------\n");
-    eprintf("-------------------------------------------------------------------------------------\n");
-
-    in_main_loop = True;
-
-    AsyncSocketStart();
-
-    for(;;)
-    {
-        if (timers == NULL)
-            ms = 500;
-        else
-        {
-            ms = timers->time - GetMilliCount();
-            if (ms <= 0)
-                ms = 0;
-
-            if (ms > 500)
-                ms = 500;
-        }
-
-        if (MsgWaitForMultipleObjects(0,NULL,0,(DWORD)ms,QS_ALLINPUT) == WAIT_OBJECT_0)
-        {
-            while (PeekMessage(&msg,NULL,0,0,PM_REMOVE))
-            {
-                if (msg.message == WM_QUIT)
-                {
-                    lprintf("ServiceTimers shutting down the server\n");
-                    return;
-                }
-
-                switch (msg.message)
-                {
-                case WM_BLAK_MAIN_READ :
-                    EnterServerLock();
-
-                    PollSession(msg.lParam);
-                    TimerActivate();
-
-                    LeaveServerLock();
-                    break;
-
-                case WM_BLAK_MAIN_RECALIBRATE :
-                    /* new soonest timer, so we should recalculate our time left...
-                    so we just need to restart the loop! */
-                    break;
-
-                case WM_BLAK_MAIN_DELETE_ACCOUNT :
-                    EnterServerLock();
-                    DeleteAccountAndAssociatedUsersByID(msg.lParam);
-                    LeaveServerLock();
-                    break;
-
-                case WM_BLAK_MAIN_VERIFIED_LOGIN :
-                    EnterServerLock();
-                    VerifiedLoginSession(msg.lParam);
-                    LeaveServerLock();
-                    break;
-
-                case WM_BLAK_MAIN_LOAD_GAME :
-                    EnterServerLock();
-                    LoadFromKod(msg.lParam);
-                    LeaveServerLock();
-                    break;
-
-                default :
-                    dprintf("ServiceTimers got unknown message %i\n",msg.message);
-                    break;
-                }
-            }
-        }
-        else
-        {
-            /* a Blakod timer is ready to go */
-
-            EnterServerLock();
-            PollSessions(); /* really just need to check session timers */
-            TimerActivate();
-            LeaveServerLock();
-        }
-    }
-#endif
 }
 
 timer_node * GetTimerByID(int timer_id)
